separate tests in `tests.rs`
