+++ pallets/simple-event/src/tests.rs
@@ -1,0 +1,91 @@
+use crate::{Module, Trait};
+use primitives::H256;
+use runtime_io;
+use runtime_primitives::{
+	testing::Header,
+	traits::{BlakeTwo256, IdentityLookup},
+	Perbill,
+};
+use support::{assert_ok, impl_outer_event, impl_outer_origin, parameter_types};
+use system::{EventRecord, Phase};
+
+// Implement the outer environment: Origin, TestEvent
+impl_outer_origin! {
+	pub enum Origin for TestRuntime {}
+}
+
+mod simple_event {
+	pub use crate::Event;
+}
+
+impl_outer_event! {
+	pub enum TestEvent for TestRuntime {
+		simple_event,
+	}
+}
+
+// Workaround for https://github.com/rust-lang/rust/issues/26925 . Remove when sorted.
+#[derive(Clone, PartialEq, Eq, Debug)]
+pub struct TestRuntime;
+
+parameter_types! {
+	pub const BlockHashCount: u64 = 250;
+	pub const MaximumBlockWeight: u32 = 1024;
+	pub const MaximumBlockLength: u32 = 2 * 1024;
+	pub const AvailableBlockRatio: Perbill = Perbill::one();
+}
+
+// The TestRuntime implements two pallet/frame traits: system, and simple_event
+impl system::Trait for TestRuntime {
+	type Origin = Origin;
+	type Index = u64;
+	type Call = ();
+	type BlockNumber = u64;
+	type Hash = H256;
+	type Hashing = BlakeTwo256;
+	type AccountId = u64;
+	type Lookup = IdentityLookup<Self::AccountId>;
+	type Header = Header;
+	type Event = TestEvent;
+	type BlockHashCount = BlockHashCount;
+	type MaximumBlockWeight = MaximumBlockWeight;
+	type MaximumBlockLength = MaximumBlockLength;
+	type AvailableBlockRatio = AvailableBlockRatio;
+	type Version = ();
+	type ModuleToIndex = ();
+}
+
+impl Trait for TestRuntime {
+	type Event = TestEvent;
+}
+
+// Hook up modules so we can make dispatched calls later
+pub type System = system::Module<TestRuntime>;
+pub type SimpleEvent = Module<TestRuntime>;
+
+pub struct ExtBuilder;
+
+impl ExtBuilder {
+	pub fn build() -> runtime_io::TestExternalities {
+		let storage = system::GenesisConfig::default()
+			.build_storage::<TestRuntime>()
+			.unwrap();
+		runtime_io::TestExternalities::from(storage)
+	}
+}
+
+#[test]
+fn test() {
+	ExtBuilder::build().execute_with(|| {
+		assert_ok!(SimpleEvent::do_something(Origin::signed(1), 32));
+
+		assert_eq!(
+			System::events(),
+			vec![EventRecord {
+				phase: Phase::ApplyExtrinsic(0),
+				event: TestEvent::simple_event(crate::Event::EmitInput(32)),
+				topics: vec![],
+			}]
+		);
+	})
+}
